Fix Visual Studio warnings.

+++ code/sky.c
@@ -75,7 +75,7 @@ void draw_sky()
     if (hour >= sunrise && hour <= sunset) {
         double t = inverse_lerp_int(sunrise * 60, sunset * 60, (hour * 60) + minute);
         sun->x = -sun->width + (int)((t) * (double)(render_width + sun->width * 2)) - (sun->width / 2);
-        sun->y = render_height - (sin(t * M_PI) * (double)render_height * 0.9); 
+        sun->y = render_height - (int)(sin(t * M_PI) * (double)render_height * 0.9); 
     }
 
     // Sky Background
